+++ entities/params/campaign.go
@@ -12,3 +12,15 @@ func (p *Campaign) TrimSpaces() {
 	p.Name = strings.TrimSpace(p.Name)
 	p.TemplateName = strings.TrimSpace(p.TemplateName)
 }
+
+// SendCampaign represents request body for POST /api/campaigns/id/start
+type SendCampaign struct {
+	Ids                 []int64           `form:"segment_id[]" validate:"gt=0,dive,required"`
+	Source              string            `form:"source" validate:"required,email,max=191"`
+	FromName            string            `form:"from_name" validate:"required,max=191"`
+	DefaultTemplateData map[string]string `form:"default_template_data" validate:"dive,required"`
+}
+
+func (p *SendCampaign) TrimSpaces() {
+	p.FromName = strings.TrimSpace(p.FromName)
+}
